Natural-voice prompts (questions v3, resume v2), merged resume data, smarter rate-limit handling
- router tries a provider's fallback model before parking it and waits for the soonest
  cooldown instead of failing when every provider is rate limited
- compact profile text and shorter descriptions for scoring (token budgets on free tiers)
- years_of_experience is a float

+++ backend/app/models/profile.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 
 from datetime import datetime, timezone
 
-from sqlalchemy import JSON, Integer, String, Text
+from sqlalchemy import JSON, Float, Integer, String, Text
 from sqlalchemy.orm import Mapped, mapped_column
 
 from ..database import Base
@@ -34,7 +34,7 @@ class Profile(Base):
     linkedin_url: Mapped[str] = mapped_column(String(300), default="")
     github_url: Mapped[str] = mapped_column(String(300), default="")
     portfolio_url: Mapped[str] = mapped_column(String(300), default="")
-    years_of_experience: Mapped[float] = mapped_column(Integer, default=0)
+    years_of_experience: Mapped[float] = mapped_column(Float, default=0)
     notice_period: Mapped[str] = mapped_column(String(100), default="")
     expected_salary: Mapped[str] = mapped_column(String(100), default="")
     current_salary: Mapped[int | None] = mapped_column(Integer, nullable=True)          # annual, in salary_currency
